// src/schemas/HeroSectionProps.ts
export interface HeroSectionProps {
    title: string;
    titleGradientText?: string;
    titleEndText?: string;
    description: string;
    contentHtml?: string;
    btnGradientIsEnabled: boolean;
    btnGradientHref: string;
    btnGradientText: string;
    btnDefaultIsEnabled: boolean;
    btnDefaultHref: string;
    btnDefaultText: string;
}